Refactor interactivity tests to make them work
- Demote test of cluster sizes from a test of plot_interactive_mapper_graph to one of the static version
- Create a new test_visualization_interactive file with a single test for plot_interactive_mapper_graph
